feat: 移除失效的下载功能
