+++ components.d.ts
@@ -7,17 +7,8 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
-    DownloadingList: typeof import('./src/components/DownloadComponents/DownloadingList.vue')['default']
-    DownloadPane: typeof import('./src/components/DownloadPane.vue')['default']
-    EpisodePane: typeof import('./src/components/DownloadComponents/EpisodePane.vue')['default']
-    MangaDirIndicator: typeof import('./src/components/WatermarkComponents/MangaDirIndicator.vue')['default']
+    MangaDirIndicator: typeof import('./src/components/MangaDirIndicator.vue')['default']
     NButton: typeof import('naive-ui')['NButton']
-    NCard: typeof import('naive-ui')['NCard']
-    NCheckbox: typeof import('naive-ui')['NCheckbox']
-    NCheckboxGroup: typeof import('naive-ui')['NCheckboxGroup']
-    NDivider: typeof import('naive-ui')['NDivider']
-    NDropdown: typeof import('naive-ui')['NDropdown']
-    NEmpty: typeof import('naive-ui')['NEmpty']
     NInput: typeof import('naive-ui')['NInput']
     NMessageProvider: typeof import('naive-ui')['NMessageProvider']
     NModal: typeof import('naive-ui')['NModal']
@@ -28,15 +19,8 @@ declare module 'vue' {
     NRadioGroup: typeof import('naive-ui')['NRadioGroup']
     NSpace: typeof import('naive-ui')['NSpace']
     NSwitch: typeof import('naive-ui')['NSwitch']
-    NTabPane: typeof import('naive-ui')['NTabPane']
-    NTabs: typeof import('naive-ui')['NTabs']
-    NText: typeof import('naive-ui')['NText']
     NTooltip: typeof import('naive-ui')['NTooltip']
-    QrCodeViewer: typeof import('./src/components/DownloadComponents/QrCodeViewer.vue')['default']
-    RemoveProgress: typeof import('./src/components/WatermarkComponents/RemoveProgress.vue')['default']
-    SearchPane: typeof import('./src/components/DownloadComponents/SearchPane.vue')['default']
-    SearchResult: typeof import('./src/components/DownloadComponents/SearchResult.vue')['default']
-    WatermarkCropper: typeof import('./src/components/WatermarkComponents/WatermarkCropper.vue')['default']
-    WatermarkPane: typeof import('./src/components/WatermarkPane.vue')['default']
+    RemoveProgress: typeof import('./src/components/RemoveProgress.vue')['default']
+    WatermarkCropper: typeof import('./src/components/WatermarkCropper.vue')['default']
   }
 }
