--- src-tauri/src/responses/cookie_status_data.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-use specta::Type;
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct CookieStatusData {
-    pub is_login: bool,
-}
