--- src-tauri/src/responses/image_token_data.rs
@@ -1,10 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-pub type ImageTokenData = Vec<UrlToken>;
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UrlToken {
-    pub url: String,
-    pub token: String,
-}
