+++ src-tauri/src/types.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
-use base64::Engine;
 use base64::engine::general_purpose;
+use base64::Engine;
 use serde::{Deserialize, Serialize};
 use specta::Type;
 
@@ -55,31 +55,4 @@ impl JpgImageData {
 pub enum ImageFormat {
     Jpeg,
     Png,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, Type)]
-pub enum QrCodeStatus {
-    NotScan,
-    Scanning,
-    Complete(String),
-    Invalid,
-    Unknown,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, Type)]
-pub struct QrCodeData {
-    pub base64: String,
-    #[serde(rename = "qrcodeKey")]
-    pub qrcode_key: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct Episode {
-    pub ep_id: i64,
-    pub ep_title: String,
-    pub comic_id: i64,
-    pub comic_title: String,
-    pub is_locked: bool,
-    pub is_downloaded: bool,
 }
