+++ src/bindings.ts
@@ -69,54 +69,6 @@ async saveConfig(config: Config) : Promise<Result<CommandResponse<null>, Command
     if(e instanceof Error) throw e;
     else return { status: "error", error: e  as any };
 }
-},
-async searchManga(keyword: string) : Promise<Result<CommandResponse<SearchData>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("search_manga", { keyword }) };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
-},
-async generateQrCode() : Promise<Result<CommandResponse<QrCodeData>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("generate_qr_code") };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
-},
-async getQrCodeStatusData(qrcodeKey: string) : Promise<Result<CommandResponse<QrCodeStatusData>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("get_qr_code_status_data", { qrcodeKey }) };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
-},
-async getBiliCookieStatusData(biliCookie: string) : Promise<Result<CommandResponse<CookieStatusData>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("get_bili_cookie_status_data", { biliCookie }) };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
-},
-async downloadEpisodes(episodes: Episode[]) : Promise<Result<CommandResponse<null>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("download_episodes", { episodes }) };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
-},
-async getMangaEpisodes(id: number) : Promise<Result<CommandResponse<Episode[]>, CommandError>> {
-    try {
-    return { status: "ok", data: await TAURI_INVOKE("get_manga_episodes", { id }) };
-} catch (e) {
-    if(e instanceof Error) throw e;
-    else return { status: "error", error: e  as any };
-}
 }
 }
 
@@ -124,29 +76,15 @@ async getMangaEpisodes(id: number) : Promise<Result<CommandResponse<Episode[]>, 
 
 
 export const events = __makeEvents__<{
-downloadEpisodeEndEvent: DownloadEpisodeEndEvent,
-downloadEpisodePendingEvent: DownloadEpisodePendingEvent,
-downloadEpisodeStartEvent: DownloadEpisodeStartEvent,
-downloadImageErrorEvent: DownloadImageErrorEvent,
-downloadImageSuccessEvent: DownloadImageSuccessEvent,
-downloadSpeedEvent: DownloadSpeedEvent,
 removeWatermarkEndEvent: RemoveWatermarkEndEvent,
 removeWatermarkErrorEvent: RemoveWatermarkErrorEvent,
 removeWatermarkStartEvent: RemoveWatermarkStartEvent,
-removeWatermarkSuccessEvent: RemoveWatermarkSuccessEvent,
-updateOverallDownloadProgressEvent: UpdateOverallDownloadProgressEvent
+removeWatermarkSuccessEvent: RemoveWatermarkSuccessEvent
 }>({
-downloadEpisodeEndEvent: "download-episode-end-event",
-downloadEpisodePendingEvent: "download-episode-pending-event",
-downloadEpisodeStartEvent: "download-episode-start-event",
-downloadImageErrorEvent: "download-image-error-event",
-downloadImageSuccessEvent: "download-image-success-event",
-downloadSpeedEvent: "download-speed-event",
 removeWatermarkEndEvent: "remove-watermark-end-event",
 removeWatermarkErrorEvent: "remove-watermark-error-event",
 removeWatermarkStartEvent: "remove-watermark-start-event",
-removeWatermarkSuccessEvent: "remove-watermark-success-event",
-updateOverallDownloadProgressEvent: "update-overall-download-progress-event"
+removeWatermarkSuccessEvent: "remove-watermark-success-event"
 })
 
 /** user-defined constants **/
@@ -155,32 +93,13 @@ updateOverallDownloadProgressEvent: "update-overall-download-progress-event"
 
 /** user-defined types **/
 
-export type Banner = { icon: string; title: string; url: string }
 export type CommandError = string
 export type CommandResponse<T> = { code: number; msg: string; data: T }
-export type Config = { outputDir: string; outputFormat: ImageFormat; outputOptimize: boolean; biliCookie: string }
-export type CookieStatusData = { isLogin: boolean }
-export type DownloadEpisodeEndEvent = DownloadEpisodeEndEventPayload
-export type DownloadEpisodeEndEventPayload = { epId: number; errMsg: string | null }
-export type DownloadEpisodePendingEvent = DownloadEpisodePendingEventPayload
-export type DownloadEpisodePendingEventPayload = { epId: number; title: string }
-export type DownloadEpisodeStartEvent = DownloadEpisodeStartEventPayload
-export type DownloadEpisodeStartEventPayload = { epId: number; title: string; total: number }
-export type DownloadImageErrorEvent = DownloadImageErrorEventPayload
-export type DownloadImageErrorEventPayload = { epId: number; url: string; errMsg: string }
-export type DownloadImageSuccessEvent = DownloadImageSuccessEventPayload
-export type DownloadImageSuccessEventPayload = { epId: number; url: string; current: number }
-export type DownloadSpeedEvent = DownloadSpeedEventPayload
-export type DownloadSpeedEventPayload = { speed: string }
-export type Episode = { epId: number; epTitle: string; comicId: number; comicTitle: string; isLocked: boolean; isDownloaded: boolean }
+export type Config = { outputDir: string; outputFormat: ImageFormat; outputOptimize: boolean }
 export type ImageFormat = "Jpeg" | "Png"
 export type JpgImageData = { info: JpgImageInfo; base64: string }
 export type JpgImageInfo = { width: number; height: number; path: string }
-export type List = { id: number; title: string; org_title: string; horizontal_cover: string; square_cover: string; vertical_cover: string; author_name: string[]; styles: string[]; is_finish: number; allow_wait_free: boolean; discount_type: number; type: number; wiki: Wiki; numbers: number; jump_value: string; real_title: string }
 export type MangaDirData = { width: number; height: number; count: number; blackBackground: JpgImageData | null; whiteBackground: JpgImageData | null }
-export type QrCodeData = { base64: string; qrcodeKey: string }
-export type QrCodeStatusData = { url: string; refresh_token: string; timestamp: number; code: number; message: string }
-export type Recommend = { id: number; title: string; horizontal_cover: string; square_cover: string; vertical_cover: string; last_short_title: string; recommendation: string; is_finish: number; total: number; allow_wait_free: boolean; author_name: string[]; styles: string[]; discount_type: number }
 export type RectData = { left: number; top: number; right: number; bottom: number }
 export type RemoveWatermarkEndEvent = RemoveWatermarkEndEventPayload
 export type RemoveWatermarkEndEventPayload = { dirPath: string }
@@ -190,10 +109,6 @@ export type RemoveWatermarkStartEvent = RemoveWatermarkStartEventPayload
 export type RemoveWatermarkStartEventPayload = { dirPath: string; total: number }
 export type RemoveWatermarkSuccessEvent = RemoveWatermarkSuccessEventPayload
 export type RemoveWatermarkSuccessEventPayload = { dirPath: string; imgPath: string; current: number }
-export type SearchData = { list: List[]; total_page: number; total_num: number; recommends: Recommend[]; similar: string; se_id: string; banner: Banner }
-export type UpdateOverallDownloadProgressEvent = UpdateOverallDownloadProgressEventPayload
-export type UpdateOverallDownloadProgressEventPayload = { downloadedImageCount: number; totalImageCount: number; percentage: number }
-export type Wiki = { id: number; title: string; origin_title: string; vertical_cover: string; producer: string; author_name: string[]; publish_time: string; frequency: string }
 
 /** tauri-specta globals **/
