+++ src/components/RemoveProgress.vue
